feat: removed old my-expenses page (deprecation)
