+++ src/app/core/mock-data/expense-filters.data.ts
@@ -1,4 +1,4 @@
-import { ExpenseFilters } from 'src/app/fyle/my-expenses/expense-filters.model';
+import { ExpenseFilters } from 'src/app/fyle/my-expenses-v2/my-expenses-filters.model';
 import { DateFilters } from 'src/app/shared/components/fy-filters/date-filters.enum';
 import { ExpenseType } from '../enums/expense-type.enum';
 import { FilterState } from '../enums/filter-state.enum';
