+++ src/app/fyle/my-expenses-v2/my-expenses-filters.model.ts
@@ -23,3 +23,13 @@ export type ReportFilters = Partial<{
   cardNumbers: string[];
   splitExpense: string;
 }>;
+
+export interface ExpenseFilters extends Omit<Filters, 'state'> {
+  state: string | string[];
+  cardNumbers: string[];
+  splitExpense: string;
+  tx_receipt_required: string;
+  tx_policy_flag: string;
+  tx_policy_amount: string;
+  or: string;
+}
