--- src/app/fyle/my-expenses/add-txn-to-report-dialog/add-txn-to-report-dialog.component.html
@@ -1,64 +1,0 @@
-<ion-header>
-  <div class="report-list--header">
-    <ion-grid>
-      <ion-row>
-        <ion-col size="1" class="report-list--header--row-icon-container">
-          <ion-icon
-            (click)="closeAddToReportModal()"
-            src="../../../../assets/svg/cross.svg"
-            class="report-list--header--row-icon"
-          ></ion-icon>
-        </ion-col>
-        <ion-col size="10" class="text-center">Add to Report</ion-col>
-        <ion-col size="1" class="report-list--header--row-icon-container">
-          <ion-icon
-            data-testid="addIcon"
-            (click)="onClickCreateReportTask()"
-            src="../../../../assets/svg/plus-square.svg"
-            class="report-list--header--row-icon"
-          ></ion-icon>
-        </ion-col>
-      </ion-row>
-    </ion-grid>
-  </div>
-</ion-header>
-<div class="fy-modal">
-  <div *ngIf="data.openReports.length; else zeroState" class="report-list--container">
-    <div class="report-list--list" *ngFor="let report of data.openReports; index as i">
-      <div data-testid="report" mat-ripple (click)="addTransactionToReport(report)">
-        <div *ngIf="i > 0" class="report-list--divider"></div>
-        <ion-grid class="ion-no-padding">
-          <ion-row>
-            <ion-col size="8.5" class="ion-no-padding">
-              <div class="report-list--purpose">{{ report.rp_purpose }}</div>
-              <div class="report-list--count">
-                {{ report.rp_num_transactions }} Expense{{ report.rp_num_transactions > 1 ? 's' : '' }}
-              </div>
-            </ion-col>
-            <ion-col size="3.5" class="ion-no-padding ion-text-right">
-              <div class="report-list--currency-amount-container">
-                <span class="report-list--currency">{{ reportCurrencySymbol }}</span>
-                <span class="report-list--amount">{{
-                  report.rp_amount || 0 | humanizeCurrency : report.rp_currency : true
-                }}</span>
-              </div>
-              <div class="ion-text-right">
-                <div class="text-center report-list--state state-pill state-{{ report.rp_state | reportState }}">
-                  {{ report.rp_state | reportState : data.isNewReportsFlowEnabled | snakeCaseToSpaceCase | titlecase }}
-                </div>
-              </div>
-            </ion-col>
-          </ion-row>
-        </ion-grid>
-      </div>
-    </div>
-  </div>
-  <ng-template #zeroState>
-    <app-fy-zero-state
-      image="../../../../../assets/images/zero-states/expenses.png"
-      header="No Reports found"
-      [taskImageStyle]="{ 'max-width': '220px', 'max-height': '124px', width: '220px', height: '124px' }"
-      message='To create a draft report please click on <ion-icon class="report-list--zero-state__icon" slot="icon-only" src="../../../../../assets/svg/plus-square.svg"></ion-icon>'
-    ></app-fy-zero-state>
-  </ng-template>
-</div>
