--- src/app/fyle/my-expenses/my-expenses-filters.model.ts
@@ -1,25 +1,0 @@
-export type Filters = Partial<{
-  state: string[];
-  date: string;
-  customDateStart: Date;
-  customDateEnd: Date;
-  receiptsAttached: string;
-  type: string[];
-  sortParam: string;
-  sortDir: string;
-  cardNumbers: string[];
-  splitExpense: string;
-}>;
-
-export type ReportFilters = Partial<{
-  state: string | string[];
-  date: string;
-  customDateStart: Date;
-  customDateEnd: Date;
-  receiptsAttached: string;
-  type: string[];
-  sortParam: string;
-  sortDir: string;
-  cardNumbers: string[];
-  splitExpense: string;
-}>;
